Week 6 exercise 10, add admin status and check for delete routes

+++ app/helpers/application_helper.rb
@@ -3,6 +3,8 @@ module ApplicationHelper
     return if current_user.nil?
 
     edit = link_to('Edit', url_for([:edit, item]), class: 'btn btn-primary')
+    return raw(edit) if !current_user.admin
+
     del = link_to('Destroy', item, method: :delete,
                                    data: { confirm: 'Are you sure?' },
                                    class: 'btn btn-danger')
